requirements.txt
# Core dependencies
flask>=3.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
werkzeug>=3.0.1

# Security
flask-cors>=4.0.0
flask-limiter>=3.5.0

# PDF processing
pypdf>=4.0.0

# LangChain ecosystem
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0

# ML/Embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0

# Utilities
numpy>=1.26.4
requests>=2.31.0